Move old test suite out of the way
The spec suite is woefully out of date, unfortunately, and the team has
decided that we need to just start over.

Typically I would delete the specs altogether, since they are still
available in git.  However, we probably want to review the older specs
for reference and reuse what we can.  To make that easier, I'm just
moving the old suite to a different directory.